Change SQL dialect substitutions to at symbol
The caret symbol was a holdover from a period when the at symbol was
not available.  But it is avaliable now again.

While it's not clear exactly what will come of the SQL dialect in terms
of its merging with query, it will certainly not be the caret symbol for
substitutions.  So it was changed in the odbc module.