Use "unsafe" injection in DUMP-TABLE for name
The parameterized query method in ODBC which puts a ? in spots
where you are going to put a variable does not work for table names.

https://stackoverflow.com/q/1208442/

The ^META variables are now being used in the SQL dialect for this
direct injection.  This is unsafe in general practice, so some ideas will
be needed to help make it less unsafe (e.g. SQL-EXECUTE/UNSAFE in
order to use the feature.)