Use SYS.UTIL.RESCUE vs. TRAP
Errors in ODBC-EXECUTE are not definitional, e.g. they can only
be trapped by the "catches everything including typos" kind of
operation.  This is put out of the way under an elaborate name
of SYS.UTIL.RESCUE to encourage those using it to think twice,
as it's not a good thing to do most of the time.

Vetting the errors in ODBC-EXECUTE and piping them around is
not a high priority task at this time.  So this just uses the
SYS.UTIL.RESCUE--which is in practice, no worse than things
were in Rebol2/Red/R3-Alpha.

But it should be addressed at some point if ODBC raises in
priority.