Best Guesses for SPREAD/ONLY Changes
This changes the .reb files (presumed to be Ren-C) to the new
conventions where /ONLY is the default, and you have to SPREAD
a block to splice it.

Some cases are easier to guess (e.g. if the right hand side
is a REDUCE and there was no /ONLY, it would have spliced a
block and would need a SPREAD now).  Other cases of just generic
APPEND X Y are harder, but this gives it a shot and then it
can be debugged later.