Honor NULL-or-# state of argless refinement
At the time of writing, you have to set an argument-less
refinement to either # or NULL.  This precludes being able
to use void, which is the result of a failed branch.

This is something that can probably be changed to allow a
void input on a refinement to be turned into null.  But
ramifications have to be considered.